move regu_var.h structs to xasl.h,access_spec.hpp, method_def.hpp
