+++ src/parser/keyword.c
@@ -233,8 +233,8 @@ static KEYWORD_RECORD keywords[] = {
   {INDEXES, "INDEXES", 1},
   {INDICATOR, "INDICATOR", 0},
   {INFINITE_, "INFINITE", 1},
-  {INF_LE, "INF_LE", 1},
-  {INF_LT, "INF_LT", 1},
+  {INF_LE_, "INF_LE", 1},
+  {INF_LT_, "INF_LT", 1},
   {INHERIT, "INHERIT", 0},
   {INITIALLY, "INITIALLY", 0},
   {INNER, "INNER", 0},
