+++ src/query/query_method.c
@@ -43,6 +43,7 @@ static int method_initialize_vacomm_buffer (VACOMM_BUFFER * vacomm_buffer, unsig
 static void method_clear_vacomm_buffer (VACOMM_BUFFER * vacomm_buffer);
 static int method_send_value_to_server (DB_VALUE * dbval, VACOMM_BUFFER * vacomm_buffer);
 static int method_send_eof_to_server (VACOMM_BUFFER * vacomm_buffer);
+static void methid_sig_freemem (method_sig_node * meth_sig);
 
 /*
  * method_clear_vacomm_buffer () - Clears the comm buffer
@@ -491,5 +492,42 @@ end:
     {
       method_clear_vacomm_buffer (&vacomm_buffer);
       return ER_FAILED;
+    }
+}
+
+/*
+ * methid_sig_freemem () -
+ *   return:
+ *   method_sig(in)     : pointer to a method_sig
+ *
+ * Note: Free function for METHOD_SIG using free_and_init.
+ */
+static void
+methid_sig_freemem (method_sig_node * method_sig)
+{
+  if (method_sig != NULL)
+    {
+      methid_sig_freemem (method_sig->next);
+      db_private_free_and_init (NULL, method_sig->method_name);
+      db_private_free_and_init (NULL, method_sig->class_name);
+      db_private_free_and_init (NULL, method_sig->method_arg_pos);
+      db_private_free_and_init (NULL, method_sig);
+    }
+}
+
+/*
+ * method_sig_list_freemem () -
+ *   return:
+ *   meth_sig_list(in)        : pointer to a meth_sig_list
+ *
+ * Note: Free function for METHOD_SIG_LIST using free_and_init.
+ */
+void
+method_sig_list_freemem (method_sig_list * meth_sig_list)
+{
+  if (meth_sig_list != NULL)
+    {
+      methid_sig_freemem (meth_sig_list->method_sig);
+      db_private_free_and_init (NULL, meth_sig_list);
     }
 }
