+++ src/query/query_method.h
@@ -53,4 +53,6 @@ extern int method_send_error_to_server (unsigned int rc, char *host, char *serve
 extern int method_invoke_for_server (unsigned int rc, char *host, char *server_name, qfile_list_id * list_id,
 				     method_sig_list * method_sig_list);
 
+void method_sig_list_freemem (method_sig_list * meth_sig_list);
+
 #endif /* _QUERY_METHOD_H_ */
